collapse native header support files
