--- packages/app/navigation/nativeHeaderIcons.ts
@@ -1,11 +1,0 @@
-export const nativeHeaderIcons = {
-  add: { uri: 'TlonHeaderAdd' },
-  back: { uri: 'TlonHeaderBack' },
-  editList: { uri: 'TlonHeaderEditList' },
-  invite: { uri: 'TlonHeaderInvite' },
-  overflow: { uri: 'TlonHeaderOverflow' },
-  refresh: { uri: 'TlonHeaderRefresh' },
-  rightSidebar: { uri: 'TlonHeaderRightSidebar' },
-  search: { uri: 'TlonHeaderSearch' },
-  settings: { uri: 'TlonHeaderSettings' },
-} as const;
