+++ packages/app/ui/components/ScreenHeaderPrimitives.tsx
@@ -1,6 +1,20 @@
-import { Icon, Text, Pressable as TlonPressable } from '@tloncorp/ui';
-import { PropsWithChildren, ReactNode } from 'react';
-import { ColorTokens, XStack, styled } from 'tamagui';
+import {
+  Button,
+  Icon,
+  Text,
+  Pressable as TlonPressable,
+  useIsWindowNarrow,
+} from '@tloncorp/ui';
+import { ComponentProps, PropsWithChildren, forwardRef, useState } from 'react';
+import { ColorTokens, TamaguiElement, XStack, styled } from 'tamagui';
+
+import { ActionSheet } from './ActionSheet';
+import {
+  type ScreenHeaderAction,
+  type ScreenHeaderIconName,
+  type ScreenHeaderMenuAction,
+  visibleScreenHeaderActions,
+} from './screenHeaderItemModel';
 
 export const HeaderIconButton = styled(Icon, {
   customSize: ['$3xl', '$2xl'],
@@ -77,14 +91,106 @@ export const HeaderControls = styled(XStack, {
   } as const,
 });
 
-export type HeaderControlProps = {
-  children?: ReactNode;
-  color?: ColorTokens;
-  disabled?: boolean;
-  onPress?: () => void;
-  side?: 'left' | 'right';
-  testID?: string;
-  type?: string;
-  accessibilityLabel?: string;
-  'aria-label'?: string;
-};
+/** React renderer for the shared item model, used by web and Android. */
+export function ScreenHeaderItemElements({
+  actions,
+  nativeHeader = false,
+}: {
+  actions: ScreenHeaderAction[];
+  nativeHeader?: boolean;
+}) {
+  const visible = visibleScreenHeaderActions(actions);
+  if (visible.length === 0) {
+    return null;
+  }
+
+  return (
+    <XStack
+      alignItems="center"
+      height={nativeHeader ? '$4xl' : undefined}
+      gap={nativeHeader ? '$l' : undefined}
+    >
+      {visible.map((action) => {
+        if (action.kind === 'menu') {
+          return <HeaderItemMenu key={action.id} action={action} />;
+        }
+        if (action.kind === 'text') {
+          return (
+            <HeaderTextButton
+              key={action.id}
+              onPress={action.disabled ? undefined : action.onPress}
+              disabled={action.disabled}
+              color={(action.tint as ColorTokens) ?? '$primaryText'}
+              testID={action.testID ?? action.id}
+            >
+              {action.text}
+            </HeaderTextButton>
+          );
+        }
+        return (
+          <HeaderIconButton
+            key={action.id}
+            type={action.icon}
+            disabled={action.disabled}
+            onPress={action.disabled ? undefined : action.onPress}
+            color={(action.tint as ColorTokens) ?? '$primaryText'}
+            backgroundColor={
+              (action.backgroundTint as ColorTokens) ?? 'transparent'
+            }
+            testID={action.testID ?? action.id}
+            aria-label={action.label}
+          />
+        );
+      })}
+    </XStack>
+  );
+}
+
+function HeaderItemMenu({ action }: { action: ScreenHeaderMenuAction }) {
+  const [open, setOpen] = useState(false);
+  const isWindowNarrow = useIsWindowNarrow();
+
+  return (
+    <ActionSheet
+      mode={isWindowNarrow ? 'sheet' : 'popover'}
+      modal
+      open={open}
+      onOpenChange={setOpen}
+      trigger={
+        <HeaderItemMenuTrigger
+          icon={action.icon}
+          aria-label={action.label}
+          onPress={isWindowNarrow ? () => setOpen(true) : undefined}
+        />
+      }
+    >
+      <ActionSheet.Content>
+        <ActionSheet.ActionGroup accent="neutral">
+          {action.items.map((item) => (
+            <ActionSheet.Action
+              key={item.id}
+              action={{
+                title: item.label,
+                action: () => {
+                  setOpen(false);
+                  item.onPress();
+                },
+              }}
+            />
+          ))}
+        </ActionSheet.ActionGroup>
+      </ActionSheet.Content>
+    </ActionSheet>
+  );
+}
+
+const HeaderItemMenuTrigger = forwardRef<
+  TamaguiElement,
+  ComponentProps<typeof Button.Frame> & { icon: ScreenHeaderIconName }
+>(function HeaderItemMenuTrigger({ icon, ...props }, ref) {
+  return (
+    <Button.Frame ref={ref} fill="text" intent="secondary" {...props}>
+      <Icon type={icon} color="$secondaryText" />
+    </Button.Frame>
+  );
+});
