+++ packages/app/ui/components/useScreenHeader.ts
@@ -1,4 +1,4 @@
-import type { UseScreenHeaderOptions } from './useScreenHeader.types';
+import type { UseScreenHeaderOptions } from './screenHeaderItemModel';
 
 export function useScreenHeader(_options: UseScreenHeaderOptions) {
   return false;
