--- packages/app/ui/components/useScreenHeader.types.ts
@@ -1,13 +1,0 @@
-import type { ReactNode } from 'react';
-
-import type { ScreenHeaderAction } from './screenHeaderItemModel';
-
-export interface UseScreenHeaderOptions {
-  enabled: boolean;
-  title: string;
-  titleElement: ReactNode;
-  usesCustomTitle: boolean;
-  backgroundColor?: string;
-  left: ScreenHeaderAction[];
-  right: ScreenHeaderAction[];
-}
